move utils function
