+++ rdagent/log/utils/__init__.py
@@ -75,3 +75,14 @@ def get_caller_info() -> CallerInfo:
         "function": frame.f_code.co_name,  # Get the caller's function name
     }
     return info
+
+
+def extract_loopid_func_name(tag):
+    """extract loop id and function name from the tag in Message"""
+    match = re.search(r"Loop_(\d+)\.([^.]+)", tag)
+    return match.groups() if match else (None, None)
+
+def extract_evoid(tag):
+    """extract evo id from the tag in Message"""
+    match = re.search(r"\.evo_loop_(\d+)\.", tag)
+    return match.group(1) if match else None
